feat: centralize _g helper
